worker/requirements.txt
# Worker Service Requirements

# Task Queue
celery==5.3.6
redis==5.0.1

# Database
neo4j==5.15.0

# PDF Processing (Mock for Graphify)
pypdf==3.17.4
pdfplumber==0.10.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-celery==0.0.0

# Utilities
python-dotenv==1.0.0
